# [WIP] 'Precompile' an execution tree to optimize performance on large requests