--- a/src/GraphQL/Compilation/CompiledField.cs
+++ b/src/GraphQL/Compilation/CompiledField.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 using GraphQL.Language.AST;
 using GraphQL.Types;