--- a/src/GraphQL/Compilation/CompiledNode.cs
+++ b/src/GraphQL/Compilation/CompiledNode.cs
@@ -1,6 +1,4 @@
-using System;
 using System.Collections.Generic;
-using System.Text;
 using GraphQL.Types;
 
 namespace GraphQL.Compilation