--- a/src/GraphQL/Execution/DocumentExecuter.cs
+++ b/src/GraphQL/Execution/DocumentExecuter.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using GraphQL.Caching;
+using GraphQL.Compilation;
 using GraphQL.Execution;
 using GraphQL.Instrumentation;
 using GraphQL.Language.AST;
@@ -334,6 +335,9 @@ namespace GraphQL
             int? maxParallelExecutionCount,
             IServiceProvider requestServices)
         {
+            var variables = inputs == null ? null : GetVariableValues(document, schema, operation?.Variables, inputs);
+            var compiled = QueryCompilation.Compile(schema, document, variables, operation);
+
             var context = new ExecutionContext
             {
                 Document = document,
@@ -342,8 +346,9 @@ namespace GraphQL
                 UserContext = userContext,
 
                 Operation = operation,
-                Variables = inputs == null ? null : GetVariableValues(document, schema, operation?.Variables, inputs),
+                Variables = variables,
                 Fragments = document.Fragments,
+                CompiledRootNode = compiled,
                 CancellationToken = cancellationToken,
 
                 Metrics = metrics,