… src/GraphQL.ApiTests/GraphQL.approved.txt
… src/GraphQL/Execution/ExecutionHelper.cs
… src/GraphQL/Execution/ExecutionNode.cs
… src/GraphQL/Execution/ExecutionStrategy.cs