This contains the precompilation code from PR #2016 retargeted to the `develop` branch.  Thanks to @IdeaHunter for supplying this PR!

### Before

|        Method | UseCaching |         Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection |      False | 1,116.985 us | 4.4080 us | 3.9076 us | 115.2344 | 39.0625 |     - | 713.36 KB |
|          Hero |      False |    17.541 us | 0.2305 us | 0.2156 us |   1.5564 |       - |     - |   9.59 KB |
| Introspection |       True |   881.542 us | 3.6298 us | 3.3953 us | 109.3750 | 38.0859 |     - | 673.39 KB |
|          Hero |       True |     3.990 us | 0.0340 us | 0.0318 us |   0.8011 |       - |     - |   4.92 KB |

### After

|        Method | UseCaching |       Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-----------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection |      False | 956.568 us | 3.4724 us | 2.8996 us | 102.5391 | 35.1563 |     - | 632.03 KB |
|          Hero |      False |  19.071 us | 0.0882 us | 0.0825 us |   1.8311 |       - |     - |  11.24 KB |
| Introspection |       True | 722.251 us | 2.8166 us | 2.4968 us |  95.7031 | 30.2734 |     - | 592.07 KB |
|          Hero |       True |   4.979 us | 0.0330 us | 0.0257 us |   1.0681 |  0.0076 |     - |   6.57 KB |

As can be seen from the above statistics, speed and memory changes drastically for the better when processing lists.  However, extremely small queries, such as the hero query, demonstrate a performance reduction.

I believe much more optimization is possible with cached queries, as the "compiled" query can typically be cached as well.  However, compilation changes currently if there is a directive based on a variable or if a different operation is selected for the same document.  I'm sure these are challenges that can be overcome.

Currently this PR:
* Needs xml comments
* Fails some tests due to changes to federated results.  See https://github.com/graphql-dotnet/graphql-dotnet/pull/2016#discussion_r527193359

I suggest we review this PR before making changes to allow caching of the compiled representation.  However, here are some preliminary results if the compilation is cached with the document:

### With cached compilation

|        Method | UseCaching |       Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-----------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection |      False | 946.143 us | 2.8453 us | 2.6615 us | 102.5391 | 35.1563 |     - | 632.05 KB |
|          Hero |      False |  18.566 us | 0.0924 us | 0.0772 us |   1.8311 |       - |     - |  11.25 KB |
| Introspection |       True | 664.145 us | 3.4373 us | 3.2153 us |  86.9141 | 27.3438 |     - | 536.36 KB |
|          Hero |       True |   3.441 us | 0.0136 us | 0.0127 us |   0.6943 |  0.0038 |     - |   4.26 KB |

As can be seen, for cached queries, the 'hero' sample now also demonstrates a benefit, in addition to an even greater benefit for large cached queries.

Note that these changes don't actually change much code, but they are breaking changes and would need to be added during a major release, such as 4.0.