@Shane32 2021-01-17
> I propose to postpone this. We should to deal with other (more understandable) PRs first.

Fine with me.  As can be seen in the performance metrics, I do see a lot of potential here, especially when coupled with caching.