@Shane32 2021-01-19
@IdeaHunter 

> What obstacles do you have?

To test precompiled queries, we can easily change a few lines of code to add a `CompiledNode` member to `Document` which doesn't compile if already set.  You can then see the performance metrics that I listed above. However, there are a few problems with this implementation:
1. Evaluation of directives changes compliation results.  Since directive arguments can be based off of variable values, it is not certain that the compilation result may be cached.  There are two potential fixes I foresee:
    * Compile the nodes regardless of directives, adding a `Func<Variables, bool>` delegate to the node which executes to determine if the node should be included.  For nodes with directives, the delegate can process the directive arguments and return a boolean.  For nodes without directives, this can be `() => true`.
    * Do not cache queries that have directives and where those directives are based on variables.  As long as any directive values are not based on variables, the compiled query can still be cached.
2. Queries are compiled only for the operation requested. There are a few potential solutions:
    * When caching is enabled, compile all of the operations for a query.
    * Add a ConcurrentDictionary to the cache entry corresponding to each of the named operations, caching the compiled query individually when the named operation is requested.
    * Separate cache entries based on both query string and operation name.
    * Do not cache entries containing more than one operation.

In addition to the above, we need confirmation that there are no problems with federation.

> And what can be done to put in 4.0?

@sungam3r is still hoping to release 4.0 by the end of the month.  In order to get it into 4.0, we would probably need all of the problems solved and implemented "yesterday" so we have time to review, add migration notes, etc.

Note that we have been making a lot of performance related improvements for 4.0. It's probably 30% faster and consumes 50% less memory already compared to 3.x.  I think that it would be nice to get this PR into 4.0 along with the rest of the performance upgrades, especially since otherwise it would have to wait for 5.0, but it would take a bit of effort immediately.  I just don't have any more time to spend on it right now.  If you write the code, I can review it.

Finally, it must be noted that at the end of the day, even if we figure out all the issues, this PR appears to provide performance benefits to (a) all cached queries (b) uncached queries with lists, but not uncached queries without lists.  Seems fine to me.