@IdeaHunter 2021-01-19
@Shane32 

Thanks for big evaluation of the problem. 
The idea behind a compilation PR was to cache fields/types in context of single execution without any forward looking primary because of directives messing things around. 

And in general if im not mistaken we talking about 3 things:
1) Document caching - which have little to nothing to do with compilation because compilation as it is now works of the document+variables(as original code without one did)
2) Fields/Types caching (aka compilation) - which shouldnt do the difference we observing for simple queries because it does whatever code would be doing anyway even without this PR, but also allocating a little bit of memory to cache the info.
3) Caching of Fields/Types caching (aka compilation caching) - which is not sound quite right once you spell full name of it. IMHO it should rather be the caching that superseeds (1) and (2) together with some twist.

Also the objection to (3) is GC pressure which still quite the problem and adding more caching only adds more fuel to fire. 

So in the light of this discussion do you still think that caching of compilation is good idea as of this PR? Because if slowness on small queries caused by mostly GC it would be better solve the GC problem first and get overral performance improvement first, and then think how to cache it properly to give even more easier life to GC