@Shane32 2021-01-19
This PR splits the execution phase into two: compiling and executing.  Common sense says that two passes are going to be slower than one pass, requiring more objects to be allocated, unless we re-use the compiled nodes (either via lists or document caching), or unless there is some other way to gain performance.  For example, if `ExecutionNode` could become a struct.

> Because if slowness on small queries caused by mostly GC it would be better solve the GC problem first and get overral performance improvement first

Yes, ideally we would be able to demonstrate performance gains whether or not caching is enabled, and then even greater performance with caching enabled.  I have not done enough experimentation and testing to know if that is possible.

For my own website that uses a graphql back-end, all queries are predictable and therefore cached, so this PR would demonstrate performance gains as soon as we can integrate it with caching, no matter if the query returned lists or not.

Other people that expose their graphql endpoint to the public may not be able to benefit from caching, and so would not see the same results.  They would see a performance increase or decrease with this PR depending on how many queries returned lists.

> So in the light of this discussion do you still think that caching of compilation is good idea as of this PR?

Yes, I think it is a necessary part of this PR.  Integrating this with caching should not be hard to implement, and no matter how much of a performance increase we can attain with compiled queries, it should demonstrate an even better increase in performance if the compiled queries can be cached.  (Even if we do something extremely basic, like only cache compiled queries that have a single operation and do not have directives.)  I just don't have the time to work on it just now.