@sungam3r 2021-01-23
> Because if slowness on small queries caused by mostly GC it would be better solve the GC problem first and get overral performance improvement first, and then think how to cache it properly to give even more easier life to GC

The same thought is constantly running through my head.