@sungam3r 2021-01-23
> In addition to the above, we need confirmation that there are no problems with federation.

I know nothing about federation 😞