@Shane32 2021-02-06
> Do we need this?

I would like to run some more benchmarks after our latest updates.  Previously there was a definite performance benefit for large queries.  The hard part is fixing all the conflicts.....