@Shane32 2021-02-22
> @Shane32 what conclusion will you have on this PR?

I don't have time to work on it at the moment.  If the conflicts can be fixed, it's pretty easy to run performance tests and assess the benefit.