@Shane32 2022-02-23
As these changes affect execution (which has not changed much), rather than parsing/serialization (which has), I do not believe the effect will have changed substantially from my comments and performance comparisons above.  However, it would be complicated to implement, and I have no plans to attempt/complete an implementation at this time.